Using StanfordNLP's dependency parser
For now, performing dependency parsing is an optional task whose implementation is motivated by the time required to preprocess a large dataset using StanfordNLP. The idea is to fully preprocess a large dataset and store the preprocessed data in a file that can be quickly loaded multiple times skipping the preprocessing phase. If dependency information becomes necessary, there is no need to preprocess the dataset again.