Removing punctuation on filter_extraction filter() (with possible bug fix)
It was found that list.remove() was not removing some of the words that had 'upos' equal to 'PUNCT', so it became necessary to create a new list with the elements we want to keep, rather than deleting elements.
Possibly, this change corrects a bug that was not being detected by the function test: there being two elements e1 and e2, where e1=e2, the removal of e2  could lead to the removal of e1 instead of e2 because list.remove() excludes the first occurrence.