Broadly enhanced storage and data loading performance
Using again the 'pickle' package to improve performance in storing and loading data.
The storage performance has increased to hundreds of files per second and the loading performance has increased to thousands of files per second. The loading performance with the previous implementation was of a few files (5 to 15) per second.