Enhanced storage and data loading performance
The stored files contain several documents and each file occupies a maximum of about 100 MB in total. If during the preprocessing phase a file needs to grow, it will be divided.
These changes allow to load and store more documents per second because it is not necessary to create/open new files so often.