Returning weights given by classifiers to features
The prediction server is returning the weights (importances or coefficients) given by the classifiers (that give access to this data) to the features/lemmas present in the text given on the request.
At the moment, the classifiers that are supported are those that are based on trees, Naive Bayes, SVM and the SGDClassifier.