Changing documents in feature extraction instead of changing data when Excel loads
This change has the advantage that all data is maintained on the preprocessed data file, without losing any content of the Excel file, and the data can be adjusted when feature extraction is performed.
It has the disadvantage that some unnecessary preprocessing can be done but, if the data is not necessary, it should be removed from the Excel file or when feature extraction is performed.
It also has the disadvantage of making the feature extraction preparation phase take longer.
This change is particularly good for when we need to perform several experiments that require changing parameters such as the classification (reduce/change tree levels).