Creating local file instead of temporary file in 'set_docs_metadata'
The changes allow to only rename the file instead of move the file, removing the overhead necessary to move the file which is noticeable when using Docker or the program is not in the system partition.
In addition, the move operation was not taken into account when showing the progress.