Update environment.yml
The old package started not finding the "sklearn.neighbors._base" module at runtime a few days ago on Windows 10 and the same is happening on Linux. This happens because the installed version of scikit-learn has the module "sklearn.neighbors.base" instead of "sklearn.neighbors._base". The new package fixes the problem by supporting the older version of scikit-learn.